refactor: move lineDiff to utils/textdiff so it can be reused
cli exported nothing at all, so the diff rendering users already see from
`charts apply --diff` could not be reused by a TUI view, an API response, or
an external consumer — each would reimplement an LCS and drift from what the
CLI prints.

Pure code motion: lineDiff becomes textdiff.Lines, splitLines stays
unexported alongside it, and cli/apply.go + cli/charts.go call through.

TestUpgradeFlagsParse moved to cli/flags_test.go rather than travelling with
the file: it exercises cli's flag parsing and only lived in diff_test.go
incidentally.

Refs #471
